Move tests for Process.clock_gettime
